Shared memory funcionando
A memória compartilhada está funcionando e os dados estão sendo compartilhados entre a poten e a Control.

+++ src/Containers/Estecamento/poten/poten_struct.h
@@ -6,7 +6,7 @@
 typedef struct{
     road_time_t time;
     const int analogpin = 0; /*pino no qual o cabo de dados do potenciometro vai ser ligado*/
-    int value_poten_in = 0; /*valor lido pelo potenciometro*/
-    int value_poten_out = 0; /*valor depositádo na memoria compartilhada do valor do ângulo*/
+    double value_poten_in = 0; /*valor lido pelo potenciometro*/
+    double value_poten_out = 0; /*valor depositádo na memoria compartilhada do valor do ângulo*/
 }POTEN_DATA;
 #endif
